backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pypdf
chromadb
tiktoken
# LangChain v1 Stack
langchain
langchain-classic
langchain-openai
langchain-community
langchain-chroma
